chore: move `wasm32_exists` logic to `cargo_near_build::cargo_native::target`
