+++ cargo-near-build/Cargo.toml
@@ -8,3 +8,4 @@ eyre = "0.6.12"
 camino = "1.1.1"
 colored = "2.0"
 dunce = "1"
+log = "0.4"
