+++ cargo-near-build/src/cargo_native.rs
@@ -1,0 +1,1 @@
+pub mod target;
