+++ cargo-near-build/src/lib.rs
@@ -1,5 +1,6 @@
 // TODO: make mod non-pub
 pub mod fs;
 // TODO: make mod non-pub
+pub mod cargo_native;
 pub mod pretty_print;
 pub mod types;
